Moved method into API layer
