+++ source/app/blueprints/pages/manage/manage_users.py
@@ -30,9 +30,8 @@ from app.datamgmt.manage.manage_users_db import get_user_details
 from app.datamgmt.manage.manage_users_db import get_user_effective_permissions
 from app.forms import AddUserForm
 from app.iris_engine.access_control.utils import ac_get_all_access_level
-from app.iris_engine.access_control.utils import ac_current_user_has_permission
 from app.models.authorization import Permissions
-from app.blueprints.access_controls import ac_requires
+from app.blueprints.access_controls import ac_requires, ac_current_user_has_permission
 from app.blueprints.responses import response_error
 
 manage_users_blueprint = Blueprint('manage_users', __name__, template_folder='templates')
